Major reorganisation of the Pop3Store to match the other stores
